Refactor server files from JS to TS
