+++ package.json
@@ -52,8 +52,7 @@
   },
   "lint-staged": {
     "*.{js,jsx,ts,tsx}": [
-      "yarn format",
-      "yarn lint"
+      
     ]
   },
   "prettier": {
