--- server/src/bootstrap.js
@@ -1,8 +1,0 @@
-'use strict';
-
-const { beforeDelete } = require('./lifecycles');
-
-module.exports = async (params) => {
-  // Lifecycles.
-  await beforeDelete(params);
-};
