--- server/src/index.js
@@ -1,18 +1,0 @@
-'use strict';
-
-const bootstrap = require('./bootstrap');
-const config = require('./config');
-const services = require('./services');
-
-module.exports = {
-  bootstrap,
-  config,
-  // contentTypes,
-  // controllers,
-  // destroy,
-  // middlewares,
-  // policies,
-  // register,
-  // routes,
-  services,
-};
