+++ server/src/index.ts
@@ -1,0 +1,10 @@
+import bootstrap from './bootstrap';
+import config from './config';
+import services from './services';
+
+export default {
+  bootstrap,
+  config,
+  services,
+};
+
