--- server/src/lifecycles/index.js
@@ -1,7 +1,0 @@
-'use strict';
-
-const beforeDelete = require('./before-delete');
-
-module.exports = {
-  beforeDelete,
-};
