+++ server/src/lifecycles/index.ts
@@ -1,0 +1,5 @@
+import beforeDelete from './beforeDelete';
+
+export default {
+  beforeDelete,
+};
