--- server/src/services/index.js
@@ -1,9 +1,0 @@
-'use strict';
-
-const config = require('./config');
-const validation = require('./validation');
-
-module.exports = {
-  config,
-  validation,
-};
