+++ server/src/services/index.ts
@@ -1,0 +1,7 @@
+import config from './config';
+import validation from './validation';
+
+export default {
+  config,
+  validation,
+};
