--- server/src/utils/get-service.js
@@ -1,7 +1,0 @@
-'use strict';
-
-const pluginId = require('./plugin-id');
-
-const getService = (name) => strapi.plugin(pluginId).service(name);
-
-module.exports = getService;
