+++ server/src/utils/getService.ts
@@ -1,0 +1,5 @@
+import pluginId from './pluginId';
+
+const getService = (name: string) => strapi.plugin(pluginId).service(name);
+
+export default getService;
