--- server/src/utils/index.js
@@ -1,11 +1,0 @@
-'use strict';
-
-const getService = require('./get-service');
-const isProtectedEntity = require('./is-protected-entity');
-const pluginId = require('./plugin-id');
-
-module.exports = {
-  getService,
-  isProtectedEntity,
-  pluginId,
-};
