+++ server/src/utils/index.ts
@@ -1,0 +1,3 @@
+export { default as getService } from './getService';
+export { default as isProtectedEntity } from './isProtectedEntity';
+export { default as pluginId } from './pluginId';
