--- server/src/utils/plugin-id.js
@@ -1,5 +1,0 @@
-'use strict';
-
-const pluginPkg = require('../../../package.json');
-
-module.exports = pluginPkg.strapi.name;
