+++ server/src/utils/pluginId.ts
@@ -1,0 +1,3 @@
+import pluginPkg from '../../../package.json';
+
+export default pluginPkg.strapi.name;
